auth/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
pydantic-settings>=2.3.0
